# speedup accessing actions one agent at a time