### Proposed change(s)
This showed up as a non-trivial portion of time (14s / 204s) when profiling with cprofile. 

Real-life profiling with 3Dball:

```
master
2020-07-22 13:27:45 INFO [stats.py:112] 3DBall: Step: 492000. Time Elapsed: 193.277 s Mean Reward: 100.000. Std of Reward: 0.000. Training.

optimized
2020-07-22 14:56:14 INFO [stats.py:112] 3DBall: Step: 492000. Time Elapsed: 187.601 s Mean Reward: 100.000. Std of Reward: 0.000. Training.
```

### Useful links (Github issues, JIRA tickets, ML-Agents forum threads etc.)



### Types of change(s)

- [ ] Bug fix
- [ ] New feature
- [x] Code refactor
- [ ] Breaking change
- [ ] Documentation update
- [ ] Other (please describe)

### Checklist
- [ ] Added tests that prove my fix is effective or that my feature works
- [ ] Updated the [changelog](https://github.com/Unity-Technologies/ml-agents/blob/master/com.unity.ml-agents/CHANGELOG.md) (if applicable)
- [ ] Updated the [documentation](https://github.com/Unity-Technologies/ml-agents/tree/master/docs) (if applicable)
- [ ] Updated the [migration guide](https://github.com/Unity-Technologies/ml-agents/blob/master/docs/Migrating.md) (if applicable)

### Other comments